kundanika version_5/kundanika version2 .1/kundanika version2/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
PyJWT
passlib
bcrypt==3.2.0
python-multipart
google-generativeai
email-validator